% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get a listing of the SDS tutorials including where they are deployed}
\usage{
get_info(pattern = NULL)
}
\arguments{
\item{pattern}{an (optional) regex to identify specific apps

Returns a data frame giving the names of the SDStutorials and the URL where they are deployed.
If `pattern` is given, return the info for just the corresponding tutorials.}
}
\description{
Get a listing of the SDS tutorials including where they are deployed
}
\examples{
get_info("bootstrap")
}
